Re-org code
